Added new plugin ColorSheep v0.1. Plugin is made on request from user CorumAnime.
"Any chance of a command that randomly recolors sheep (permanently) within a set distance of the player? And if possible it also removes x amount of an item (eg/redstone 331) from inventory in order to do the command? That would be a super cool addition to any server then!"

That pretty much describes the plugin. Only difference is that its not included into DiscoSheep but rather as its own plugin. This was to avoid some clutter in DiscoSheep. But if you want both plugins there is nothing stopping you. 

Also updated license.txt. We are in year 2011 now. Not 2010. HAPPY NEW YEAR. 3 months too late tho...